pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-clickhouse-connect"
authors = [
    { name = "Sergey Irin", email = "serjdukareff@gmail.com" },

]
license = { text = "Apache License 2.0" }
description = "Airflow Clickhouse Provider based on http connection"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
]
dynamic = [
    "version"
]
requires-python = "~=3.12"
dependencies = [
    "apache-airflow>=2.10",
    "clickhouse-connect>=0.8.3"
]

[project.urls]
Source = "https://github.com/RinSer/airflow-provider-clickhouse-connect/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "clickhouse_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "clickhouse_provider.__version__" }

[tool.setuptools.packages.find]
exclude = [
    "*tests.*",
    "*tests"
]
